full JWT implementation

+++ Udemy 2 - REST APIs with Flask and Python in 2025/8 - JWT/resources/item.py
@@ -1,7 +1,6 @@
-import uuid
-from flask import request
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError
 
 from db import db
@@ -13,6 +12,7 @@ blp = Blueprint("items", __name__, description="Operations on items")
 
 @blp.route("/item/<string:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         # try:
@@ -22,6 +22,7 @@ class Item(MethodView):
         item = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required(fresh=True)
     def delete(self, item_id):
         # try:
         #     del items[item_id]
@@ -30,11 +31,16 @@ class Item(MethodView):
         #     abort(404, message="Item not found")
         # ----------------------
         # raise NotImplementedError("Deleting an item is not implemented")
+        jwt = get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
+
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
         return {"message": "Item deleted."}
 
+    @jwt_required() # added by me
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -64,6 +70,7 @@ class Item(MethodView):
 
 @blp.route("/item")
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         # por culpa del uso de marshmallow, ya no es un objeto/diccionario con la lista de items, sino meramente la lista y ya
@@ -72,6 +79,7 @@ class ItemList(MethodView):
         # return items.values()
         return ItemModel.query.all()
 
+    @jwt_required()
     @blp.arguments(ItemSchema)
     @blp.response(200, ItemSchema)
     def post(self, item_data):
